deep-research-agent/pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
